+++ BankingApp.Entities/Entities/DepositeHistory.cs
@@ -8,7 +8,7 @@ namespace BankingApp.Entities.Entities
         public string CalculationFormula { get; set; }
         public decimal DepositeSum { get; set; }
         public int MonthsCount { get; set; }
-        public int Percents { get; set; }
+        public float Percents { get; set; }
         public DateTime CalulationDateTime { get; set; }
         public virtual IList<DepositeHistoryItem> DepositeHistoryItems { get; set; }
     }
